requirements.txt
flask
requests
streamlit
python-dotenv
